Rename matrix-synapse service to same as package
Before deploying this make sure /etc/matrix-synapse/conf.d/ is empty.

Previously the service was named synapse.service and the package also
contained one named matrix-synapse.service. Use the same name as the
package and delete the old one to avoid any confusion. Move the changed
parts of the service to a drop in file and use the packaged service.

Decrease restart time to 3 minutes, it was unrealistically long.

Also move all related parts from init.sls to synapse.sls to not have
them in two files.